Unfinished version Fixes issue where mutton would not drop from killed sheep Fixes broken fortune formula Small balance change to looting formula

+++ src/main/java/net/keegancuff/progrework/mixin/LootingFormula.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 @Mixin(LootingEnchantLootFunction.class)
 public class LootingFormula {
     @ModifyVariable(method = "process", at = @At("STORE"))
-    private double injected(float f){
+    private float injected(float f){
         return Math.min(1f, f);
     }
 }
